Move FeltSerializer to separate file
