+++ lib/src/main/kotlin/com/swmansion/starknet/data/serializers/FeltSerializer.kt
@@ -1,0 +1,24 @@
+package com.swmansion.starknet.data.serializers
+
+import com.swmansion.starknet.data.types.Felt
+import com.swmansion.starknet.extensions.toHex
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
+
+internal object FeltSerializer : KSerializer<Felt> {
+    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Felt", PrimitiveKind.STRING)
+
+    override fun deserialize(decoder: Decoder): Felt {
+        val hex = decoder.decodeString()
+
+        return Felt.fromHex(hex)
+    }
+
+    override fun serialize(encoder: Encoder, value: Felt) {
+        encoder.encodeString(value.value.toHex())
+    }
+}
