+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Felt.kt
@@ -2,14 +2,9 @@ package com.swmansion.starknet.data.types
 
 import com.swmansion.starknet.data.parseHex
 import com.swmansion.starknet.data.types.conversions.ConvertibleToCalldata
+import com.swmansion.starknet.data.serializers.FeltSerializer
 import com.swmansion.starknet.extensions.toHex
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import java.math.BigInteger
 
 @Serializable(with = FeltSerializer::class)
@@ -118,16 +113,3 @@ data class Felt(val value: BigInteger) : Comparable<Felt>, ConvertibleToCalldata
     }
 }
 
-internal object FeltSerializer : KSerializer<Felt> {
-    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Felt", PrimitiveKind.STRING)
-
-    override fun deserialize(decoder: Decoder): Felt {
-        val hex = decoder.decodeString()
-
-        return Felt.fromHex(hex)
-    }
-
-    override fun serialize(encoder: Encoder, value: Felt) {
-        encoder.encodeString(value.value.toHex())
-    }
-}
